backend/requirements.txt
### Backend
fastapi
uvicorn

### Pipeline
yfinance
transformers
torch
pandas
numpy
python-dotenv

### Models
pydantic
pydantic-settings